improved css
